+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Job/JobParameters/ConstraintCollectionProvider/CleanRemovedAttribute.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider;
+
+use Akeneo\Tool\Component\Batch\Job\JobInterface;
+use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
+use Symfony\Component\Validator\Constraints\All;
+use Symfony\Component\Validator\Constraints\Collection;
+use Symfony\Component\Validator\Constraints\NotBlank;
+use Symfony\Component\Validator\Constraints\NotNull;
+use Symfony\Component\Validator\Constraints\Type;
+
+final class CleanRemovedAttribute implements ConstraintCollectionProviderInterface
+{
+    private array $supportedJobNames;
+
+    public function __construct(array $supportedJobNames)
+    {
+        $this->supportedJobNames = $supportedJobNames;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getConstraintCollection(): Collection
+    {
+        return new Collection([
+            'fields' => [
+                'attribute_codes' => [
+                    new NotNull(),
+                    new Type('array'),
+                    new All([new Type(['type' => 'string']), new NotBlank()])
+                ],
+            ],
+        ]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supports(JobInterface $job): bool
+    {
+        return in_array($job->getName(), $this->supportedJobNames);
+    }
+}